requirements.txt
appnope
asttokens
attrs
autopep8
backcall
beautifulsoup4
bleach
debugpy
decorator
defusedxml
entrypoints
et-xmlfile
executing
fastjsonschema
importlib-resources
ipykernel
ipython
jedi
Jinja2
jsonschema
jupyter-client
jupyter-core
jupyterlab-pygments
MarkupSafe
matplotlib-inline
mistune
nbclient
nbconvert
nbformat
nest-asyncio
numpy
openpyxl
packaging
pandas
pandocfilters
parso
pexpect
pickleshare
Pillow
Pillow-PIL
pip
prompt-toolkit
psutil
ptyprocess
pure-eval
pycodestyle
Pygments
pyparsing
pyrsistent
python-dateutil
pytz
pyzmq
setuptools
six
soupsieve
stack-data
tinycss2
toml
tornado
traitlets
wcwidth
webencodings
wheel
XlsxWriter
zipp